`automateShoppingList` works, but need to test
